Fix multiple warning: comparison of integer expressions of different signedness
  When compiling for ESP32 with ESP-IDF-v4.2.3 there were multiple warning:
"comparison of integer expressions of different signedness: 'int' and 'size_t'
{aka 'unsigned int'}"

+++ src/libAtomVM/bitstring.c
@@ -63,7 +63,7 @@ bool bitstring_insert_any_integer(uint8_t *dst, avm_int_t offset, avm_int64_t va
     if (bs_flags != 0) {
         return false;
     }
-    for (int i = 0; i < n; ++i) {
+    for (uint8_t i = 0; i < n; ++i) {
         int k = (n - 1) - i;
         int bit_val = (value & (0x01 << k)) >> k;
         if (bit_val) {
